tests/test_ai_summary_migrated.py
"""
ai_summary_test的pytest版本
从根目录迁移并转换为标准pytest测试
"""

import pytest
from unittest.mock import Mock, patch

from tests.test_utils import TestDataGenerator, TestAssertions
from tests.memory import test_memory, get_project_root


class TestAisummary:
    """测试类"""
    
    def setup_method(self):
        """测试前准备"""
        self.test_data_gen = TestDataGenerator()
        self.assertions = TestAssertions()
        self.test_memory = test_memory
    
    def teardown_method(self):
        """测试后清理"""
        pass

    def test_ai_summary_to_wechat_pytest(self):
        """pytest版本的test_ai_summary_to_wechat"""
        # TODO: 实现pytest版本的测试逻辑
        pass


# 可以用于手动测试的函数
def manual_test():
    """手动测试函数"""
    print("运行手动测试...")
    # TODO: 实现手动测试逻辑
    print("手动测试完成！")


if __name__ == "__main__":
    manual_test()
