pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wechat-rss-auto-sender"
version = "0.1.0"
description = "自动获取RSS新闻并通过微信发送AI总结的服务"
authors = [{name = "fulln", email = "a114mypr1nt@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "wxauto>=3.9.8.17",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "openai>=1.3.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3"
]

[project.scripts]
wechat-rss-sender = "src.main:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",                    # 显示所有测试结果的简短摘要
    "-v",                     # 详细输出
    "--tb=short",            # 简短的错误回溯
    "--strict-markers",      # 严格标记模式
    "--cov=src",             # 代码覆盖率
    "--cov-report=term-missing",  # 显示缺失的覆盖行
    "--cov-report=html:htmlcov",  # HTML覆盖率报告
    "--disable-warnings"     # 禁用警告
]
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "api: marks tests that require API access",
    "network: marks tests that require network access",
    "wechat: marks tests related to WeChat functionality",
    "rss: marks tests related to RSS functionality",
    "ai: marks tests related to AI/summarization functionality",
    "mock: marks tests that use mocking",
    "real: marks tests that require real API credentials",
    "asyncio: marks tests as async tests"
]
# 异步测试支持
asyncio_mode = "auto"
# 测试超时
timeout = 300
# 过滤警告
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "wxauto.*",
    "feedparser.*",
    "schedule.*",
]
ignore_missing_imports = true